Make the unit of timestamp in the program refer to TIMESTAMP_UNIT_NS. (so that 1 tick can stand for 0.1 ms hence has larger range for simulation) - The timestamp unit is no longer in nano second, therefore variable names have to be modified next. - The input/output file needs to be handled to accept data in different unit. The concept of the horizontal scale factor has been changed from "divider" to "multiplier" for TRACE_HORIZONTAL_SCALE_FACTOR.

+++ src/com/illinois/rts/visualizer/TraceGroup.java
@@ -190,7 +190,7 @@ public class TraceGroup {
         }
     }
 
-    public void applyHorizontalScale(int inScale)
+    public void applyHorizontalScale(double inScale)
     {
         for (Trace currentTrace : traces)
         {
